[Feature #160806298] Cleaned up the code in the files
