+++ app/tests/v1/test_views_register.py
@@ -1,10 +1,17 @@
 import unittest
 import json
 
-from app.tests.v1.test_db import TestBase
 from app import create_app, migration
 
-class UserRegisterTestCase(TestBase):
+class UserRegisterTestCase(unittest.TestCase):
+    """Unit testiing for the user regsitration endpoint"""
+    def setUp(self):
+        """Initialize the app and database connections"""
+        self.app = create_app(config_name="testing")
+        self.client = self.app.test_client
+    
+        with self.app.app_context():
+            migration.main(config='testing')
 
     def test_register_user(self):
         response = self.client().post('/api/v2/auth/signup', data = json.dumps({
@@ -17,9 +24,4 @@ class UserRegisterTestCase(TestBase):
             "admin" : False
         }), content_type = 'application/json')
 
-        # self.assertIn('coolkid', str(response.data))
         self.assertEqual(response.status_code, 201)
-
-
-    
-
