+++ requirements.txt
@@ -26,7 +26,6 @@ mccabe==0.6.1
 more-itertools==4.3.0
 parso==0.3.1
 pickleshare==0.7.4
-pkg-resources==0.0.0
 pluggy==0.7.1
 prompt-toolkit==1.0.15
 psycopg2==2.7.5
